homelab_airflow/dags/bfs_statatlas/requirements.txt
apache-airflow
psycopg2
sqlalchemy==1.4.52
pymongo
redis
sqlmodel

pandas==2.1
requests
tqdm
